A new constant is added to this configuration file
named Time_Zone to set our time zone to be set in bootStrap file and actually to avoid getting PHP errors (notice).

And also two other constant with the name of `Default_Controller` and `Default_Method` is added to this configuration file that the first one is to set our default controller name and the second one for the default methdod (action) instead of having to use the methods that was defined for this purpuse and by adding these two extra constants, we can easily do the job, But also the old way of setting the default controller and method (action) is still working and not removed.