A private method name parseQueryString() is added to the class to parse the URL and return the value if is set, otherwise false.
Another new method name specifyingControllerMethodArguments() also is
added to the class which parses the URL given to it and sets the
controller and method and arguments if thery are not specified the
default values will be set and also takes an argument that it' value
comes form the parseQueryString() method.

And also construct method that was doeing some stuff now just calls
these twi method that we mentioned them above.

The result of this modification is because of making framework
independant of web-server and if one day wanted to support other
frameworks it will be easy to achieve it.

And also that because of this modification the style of URLs have
changed to like this: http://URL?q=controller/method/arguments

But it also works woth the old way (User friendly) that is usig apache
(htaccess) RewiteMode to do that,
	which URL looks like this: http://URL/controller/method/arguments