A new static property named $enableRewrite with the default value of false is added.
Which stores the status of the link to be QueryString like or rewrited
(using apache) if is true then it will use apache for URL rewriting for
userfriendly URL other wise the default one which is query string like
will be used.

And also a new method named go which is created that has 4 parameters
that first one takes the name of the controller and the second one takes
the name of the method and the third one takes the arguments and an
arraye or a string and the fourth one witch its default value is false
couse Framework to go to that link or return the made link (controller,
		method, ... all toghether).
If the fourth parameter is true then a URL will be returned but if is
false then it will directly redirect to the make address.