Tests folder is make so that all Tests written for framework will be placed it it.
Tests/library/FormHandler/Email.php,
	Tests/library/FormHandler/Input.php,
	Tests/library/FormHandler/index.php are three test class files which
	is written for newly developer FormHandler library.

Tests/bootStrap.php file is to include vendor/autoload.php which will be
created by composer so that phpunit can use it to find classes which is
calles in Test cases, so that we will not need to manually include those
libraries in out files using (require_once, ...).
* For more information about composer please refer to http://getcomposer.org/
website.

Tests/composer.json is for composer as kind of guideline for composer so
that composer can get understand what to do, which in fatc will create
autoloader files for us with the requested libraries (classes) in it.
Actually having this file is mandatory for composer to create
autoloader.

Tests/phpunit.xml file is for ease of use.
when calling phpunit command in Tests folder it can refer to this file
so that it can find test
files and load bootStrap file.
Then will use autoloader which composer has
made so that it can load other libraries which is called in test cases
and finally test written test cases.

*Tip: For creating autoloader file run this command 'composer install'
	 inside Tests folder so that composer can make autoloader for us in it.