Translator library refactored with some changes.
Some new methods added to this library in order to enhance the
functionality of the library for what is it doing.

Two methods setDefaultLang and setLangFileExtension is deprecated now
because of their name which was not clear in indicating that what they
are exactly doing and programer advised to user other methods for same
purposes. (language and languageFileExtention respectively ).

This library is independant of Rastgo-Framework now and its possible to
use it out of framework.

Till now languages folder was harde coded in class which it is solved
now and programer has to define where he/she has put languages (address
		of root folder) by languagesFolder method in order to get this
library work.

Test cases are written for this library and are located in Tests folder
in root of the framework.